0130-demo/backend/requirements.txt
Flask==2.1.1
psycopg2-binary
Werkzeug>=2.0.0,<2.2.0
Flask-SQLAlchemy==2.5.0
SQLAlchemy==1.3.24
Flask-Migrate==3.1.0


